pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cognitive-memory-layer"
dynamic = ["version"]
description = "Neuro-inspired memory system for LLMs (server + Python SDK)"
readme = "packages/py-cml/README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
authors = [
    { name = "CognitiveMemoryLayer Team" },
]
keywords = ["memory", "llm", "ai", "cognitive", "sdk", "rag", "event-sourcing", "cqrs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "neo4j>=5.0",
    "redis>=5.0",
    "celery>=5.4",
    "openai>=1.0",
    "sentence-transformers>=3.0",
    "tiktoken>=0.8",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "cachetools>=5.0",
    "tenacity>=8.0",
    "prometheus-client>=0.21",
    "httpx>=0.27",
    "spacy>=3.7",
    "semchunk",
    "transformers",
    "einops>=0.7.0",
]

[project.optional-dependencies]
embedded = [
    "aiosqlite>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "neo4j>=5.0",
    "redis>=5.0",
    "openai>=1.0",
    "sentence-transformers>=3.0",
    "tiktoken>=0.8",
    "structlog>=24.0",
    "pydantic-settings>=2.0",
]
gemini = ["google-generativeai"]
claude = ["anthropic"]
eval = [
    "requests>=2.32",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "openai>=1.0",
]
modeling = [
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "joblib>=1.4",
    "pyarrow>=17.0",
    "datasets>=2.20",
    "httpx>=0.27",
    "tqdm>=4.66",
]
dev = [
    "types-requests>=2.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.34",
    "black>=24.0",
    "ruff>=0.8",
    "mypy>=1.0",
    "httpx>=0.28",
    "alembic>=1.14",
    "testcontainers>=4.6",
    "psycopg2-binary>=2.9",
    "aiosqlite>=0.20",
    "pre-commit>=3.0",
]
docs = [
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.27",
]

[project.scripts]
cml-eval = "cml.eval.cli:main"
cml-models = "cml.modeling.cli:main"

[tool.hatch.metadata.hooks.custom]
path = "hatch_build.py"

[project.urls]
Homepage = "https://github.com/avinash-mall/CognitiveMemoryLayer"
Documentation = "https://github.com/avinash-mall/CognitiveMemoryLayer/tree/main/packages/py-cml#readme"
Repository = "https://github.com/avinash-mall/CognitiveMemoryLayer"
Issues = "https://github.com/avinash-mall/CognitiveMemoryLayer/issues"
Changelog = "https://github.com/avinash-mall/CognitiveMemoryLayer/blob/main/packages/py-cml/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src", "packages/py-cml/src/cml"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".git.corrupt"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH", "RUF"]
ignore = ["E501", "TC001", "TC003", "B007", "B008", "B904", "SIM102", "SIM105", "N818", "RUF006", "RUF012", "RUF034", "RUF043"]

# Import after sys.path is intentional in these paths
[tool.ruff.lint.per-file-ignores]
"evaluation/locomo_plus/scripts/*.py" = ["E402"]
"evaluation/locomo_plus/task_eval/*.py" = ["E402"]
"tests/unit/test_models_prepare_incremental.py" = ["E402"]
"migrations/env.py" = ["E402"]
"src/memory/neocortical/fact_store.py" = ["B010"]
[tool.ruff.lint.isort]
known-first-party = ["cml", "src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
exclude = "packages/py-cml/tests"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "src.api.*",
    "src.consolidation.*",
    "src.core.*",
    "src.extraction.*",
    "src.forgetting.*",
    "src.memory.*",
    "src.reconsolidation.*",
    "src.retrieval.*",
    "src.storage.*",
    "src.utils.*",
    "src.celery_app",
    "src.main",
]
ignore_errors = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "packages/py-cml/tests"]
markers = [
    "integration: integration tests (require DB, Redis, Neo4j, or external services)",
    "e2e: end-to-end tests (require full server or embedded engine)",
    "embedded: embedded mode tests (require embedded extras and engine)",
    "slow: slow tests",
    "requires_llm: marks tests that need a reachable LLM (skip when unconfigured or rate-limited)",
]
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning",
]
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src", "packages/py-cml/src/cml"]
branch = true

[tool.coverage.report]
show_missing = true
# Ramp up toward 70% as tests expand; target 70% (Issues.md A-04)
fail_under = 60
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "pass",
]
